Refactor delete_old_backups to only delete backup files

+++ apps/base/tasks.py
@@ -37,9 +37,10 @@ class BackupDatabaseTask(MoodyPeriodicTask):
 
     def delete_old_backups(self):
         for backup_file in os.listdir(settings.DATABASE_BACKUPS_PATH):
-            backup_filename = os.path.join(settings.DATABASE_BACKUPS_PATH, backup_file)
-            logger.info('Deleting old backup {}'.format(backup_filename))
-            os.unlink(backup_filename)
+            if any([backup_file.startswith(model) for model in settings.DATABASE_BACKUP_TARGETS]):
+                backup_filename = os.path.join(settings.DATABASE_BACKUPS_PATH, backup_file)
+                logger.info('Deleting old backup {}'.format(backup_filename))
+                os.unlink(backup_filename)
 
     """Task to backup mission critical database tables"""
     @update_logging_data
